requirements.txt
aniso8601==9.0.1
attrs==21.2.0
bcrypt==3.2.0
bidict==0.21.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
cryptography==36.0.0
Deprecated==1.2.13
elasticsearch==7.15.2
Flask==2.0.2
flask-restx==0.5.1
flask-sock==0.4.0
Flask-WTF==1.0.0
future==0.18.2
h11==0.12.0
idna==3.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.3.2
MarkupSafe==2.0.1
netmiko==3.4.0
ntc-templates==3.0.0
pan-python==0.16.0
paramiko==2.8.0
pkg_resources==0.0.0
pyaml==21.10.1
pycparser==2.21
PyNaCl==1.4.0
pyrsistent==0.18.0
pyserial==3.5
python-decouple==3.5
python-engineio==4.3.0
python-nmap==0.7.1
python-socketio==5.5.0
pytz==2021.3
PyYAML==6.0
redis==4.0.2
requests==2.26.0
rq==1.10.0
scp==0.14.1
simple-websocket==0.5.0
six==1.16.0
tenacity==8.0.1
textfsm==1.1.2
Turbo-Flask==0.8.0
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
wsproto==1.0.0
WTForms==3.0.0
xmltodict==0.12.0
zipp==3.6.0
